auto_scripts/api/requirements.txt
# Python UI & API Automation Framework Dependencies

# Selenium for UI automation
selenium==4.15.2

# Pytest for test execution
pytest==7.4.3
pytest-html==4.1.1
pytest-xdist==3.5.0

# Requests for API testing
requests==2.31.0

# YAML parser for configuration
PyYAML==6.0.1

# Logging and reporting
allure-pytest==2.13.2

# WebDriver Manager (optional, for automatic driver management)
webdriver-manager==4.0.1

# Additional utilities
python-dotenv==1.0.0
